Add files via upload
I cannot push the rest of my commits due to the nature of the benchmark project creating new files and folders when ran. These folders were added to my commits and because of this these commits and any future commits cannot be pushed to my repository as the size of these files are too large. I hope you understand. I am willing to come show you myself if necessary.